+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - `join` to match `Array.prototype`.
 - `cartesianProduct` to generate large sequence of products.
+- `powerSet` to generate large sequence of power sets.
 
 ### Changed